\name{plot.indice}
\alias{plot.indice}
\title{Affichage circulaire des indices de qualité de l'air}
\usage{
  plot.indice(x, y, cex = 1.7)
}
\arguments{
  \item{x}{un objet de la calsse
  \code{\link[timetools:TimeIntervalDataFrame-class]{TimeIntervalDataFrame}}
  contenant les indices à représenter}

  \item{y}{valeur unique (numérique, caractère) référençant
  la colonne à représenter}

  \item{cex}{character expansion, cf
  \code{\link[graphics]{par}}}
}
\description{
  fonction permettant de représenter des indices de qualité
  de l'air de manière rigoulotte :)
}
\seealso{
  \code{\link{indicesXR2R}}
}

